started bandnames routes and error handling
